флаг, пневмопушка

+++ code/game/objects/items/toys.dm
@@ -68,6 +68,8 @@
 	return ..()
 
 /obj/item/toy/balloon/throw_impact(atom/hit_atom, datum/thrownthing/throwingdatum)
+	if(..())
+		return
 	if(src.reagents.total_volume >= 1)
 		visible_message("<span class='warning'>The [src] bursts!</span>","You hear a pop and a splash.")
 		reagents.reaction(get_turf(hit_atom))
@@ -140,7 +142,7 @@
 
 	. = ..()
 
-/obj/item/toy/spinningtoy/throw_at(atom/target, range, speed, mob/thrower, spin = TRUE, diagonals_first = FALSE, datum/callback/callback)
+/obj/item/toy/spinningtoy/throw_at(atom/target, range, speed, mob/thrower, spin = TRUE, diagonals_first = FALSE, datum/callback/callback, can_be_catched)
 	return
 
 /obj/item/toy/spinningtoy/attack_hand(mob/user)
@@ -511,7 +513,8 @@
 	w_class = SIZE_MINUSCULE
 
 /obj/item/toy/snappop/throw_impact(atom/hit_atom, datum/thrownthing/throwingdatum)
-	..()
+	if(..())
+		return
 	var/datum/effect/effect/system/spark_spread/s = new /datum/effect/effect/system/spark_spread
 	s.set_up(3, 1, src)
 	s.start()
@@ -1023,12 +1026,14 @@ Owl & Griffin toys
 	w_class = SIZE_TINY
 
 /obj/item/toy/minimeteor/throw_impact(atom/hit_atom, datum/thrownthing/throwingdatum)
-	if(!..())
-		playsound(src, 'sound/effects/meteorimpact.ogg', VOL_EFFECTS_MASTER)
-		for(var/mob/M in orange(10, src))
-			if(M.stat == CONSCIOUS && !isAI(M))\
-				shake_camera(M, 3, 1)
-		qdel(src)
+	if(..())
+		return
+
+	playsound(src, 'sound/effects/meteorimpact.ogg', VOL_EFFECTS_MASTER)
+	for(var/mob/M in orange(10, src))
+		if(M.stat == CONSCIOUS && !isAI(M))\
+			shake_camera(M, 3, 1)
+	qdel(src)
 
 /*
  * A Deck of Cards
